defrag.f2fs: fix missing SSA updates
Previously SSA is updated if it is not included in current segment info.
But, defrag.f2fs doesn't handle current segment info during the process, and
instead lastly update the whole current segment info at a time.
So, we need to update summary entries all the time.
Otherwise, we can lose the SSA entry.

Change-Id: Ic94dc3b3b589620d30c7c442ec2b580ce4c76cad
Signed-off-by: Jaegeuk Kim <jaegeuk@kernel.org>
Signed-off-by: Hriday Sharma <hridaysharma42@gmail.com>